apps/news-crawler/requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
kafka-python==2.0.2
python-dotenv==1.0.1
